pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.51.0",
    "fitz>=0.0.1.dev2",
    "numpy>=2.2.6",
    "openai>=1.79.0",
    "opencv-python>=4.11.0.86",
    "opencv-python-headless>=4.11.0.86",
    "pillow>=11.2.1",
    "pymupdf>=1.25.5",
    "streamlit>=1.45.1",
]
